Move C++ `ExportType` to its own file
