+++ crates/c-api/include/wasmtime/types.hh
@@ -1,6 +1,7 @@
 #ifndef WASMTIME_TYPES_HH
 #define WASMTIME_TYPES_HH
 
+#include <wasmtime/types/export.hh>
 #include <wasmtime/types/func.hh>
 #include <wasmtime/types/global.hh>
 #include <wasmtime/types/import.hh>
